Fix: use of GetBindingDirPath without dynapi
For classic binding < 3 without dynapi this fix the error
at compile time about conflicting type for this function.
This worked on eel and previous version because 'struct dynapi'
was declared and available for version 2 but now since binder
is in V3 then it isn't the case so function signature is not
valid anymore.

Change-Id: I135a723d21d70b8c54f4cab1c534210757318ed0
Signed-off-by: Romain Forlot <romain.forlot@iot.bzh>