timer: Fix stopping a running timer
This fix a bug where using 1 don't stop the timer to trigger
and now TimerStop really stop the timer...
sd_event_source_unref only decrement the ref counter of the
sd_event_source. If we want to stop returning -1 in the callback
will make sure we effectively stop the timer instead of only
decrement the ref counter.

Bug-AGL: SPEC-3073

Change-Id: I4619ff1695c398a7c9683b58bd193b8c21b9abb1
Signed-off-by: Romain Forlot <romain.forlot@iot.bzh>